# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Marketplace AI API",
    "version": "1.0.0",
    "description": "API documentation for the Marketplace AI platform"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for webhook authentication"
      }
    },
    "schemas": {
...
      "Notification": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Notification ID"
          },
          "recipient": {
            "$ref": "#/components/schemas/User"
          },
          "type": {
            "type": "string",
            "enum": [
              "order_created",
              "payment_received",
              "refund_processed",
              "auction_ended",
              "rental_request",
              "system"
            ],
            "description": "Notification type"
          },
          "title": {
            "type": "string",
            "description": "Notification title"
          },
          "message": {
            "type": "string",
            "description": "Notification message"
          },
          "data": {
            "type": "object",
            "description": "Additional notification data"
          },
          "read": {
            "type": "boolean",
            "description": "Whether the notification has been read"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "description": "Notification priority"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Error status"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            },
            "description": "Validation errors"
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "Total number of items"
          },
          "page": {
            "type": "integer",
            "description": "Current page number"
          },
          "pages": {
            "type": "integer",
            "description": "Total number of pages"
          },
          "limit": {
            "type": "integer",
            "description": "Number of items per page"
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message"
          },
          "data": {
            "type": "object",
            "description": "Response data"
          }
        }
      },
...
      "ComplexResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the operation was successful"
          },
          "message": {
            "type": "string",
            "description": "Response message"
          },
          "data": {
            "type": "object",
            "description": "Response data"
          },
          "metadata": {
            "type": "object",
            "description": "Additional response metadata",
            "properties": {
              "processingTime": {
                "type": "number"
              },
              "serverTime": {
                "type": "string",
                "format": "date-time"
              },
              "pagination": {
                "$ref": "#/components/schemas/Pagination"
              },
              "filters": {
                "type": "object"
              },
              "sorting": {
                "type": "object"
              }
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "details": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/admin/notifications": {
      "post": {
        "summary": "Create a system notification",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Notification"
              },
              "examples": {
                "NotificationExample": {
                  "value": {
                    "title": "System Maintenance",
                    "message": "Scheduled maintenance at 2am UTC.",
                    "type": "system",
                    "priority": "high"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "System notification created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "SuccessResponse": {
                    "value": {
                      "message": "Notification created",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "ValidationError": {
                    "$ref": "#/components/examples/ErrorExamples/ValidationError"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthorizationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthorizationError"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notifications": {
      "get": {
        "summary": "Get user notifications",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "unread",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter for unread notifications"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string"
            },
            "description": "Filter by notification type"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Limit number of notifications"
          }
        ],
        "responses": {
          "200": {
            "description": "List of notifications",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplexResponse"
                },
                "examples": {
                  "ListResponse": {
                    "$ref": "#/components/examples/ResponseExamples/ListingCreated"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/unread/count": {
      "get": {
        "summary": "Get count of unread notifications",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Unread notification count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "SuccessResponse": {
                    "value": {
                      "count": 5
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/{notificationId}/read": {
      "patch": {
        "summary": "Mark a notification as read",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "notificationId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Notification ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Notification marked as read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "SuccessResponse": {
                    "value": {
                      "message": "Notification marked as read",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Notification not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "NotFoundError": {
                    "$ref": "#/components/examples/ErrorExamples/NotFoundError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/read/all": {
      "patch": {
        "summary": "Mark all notifications as read",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All notifications marked as read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "SuccessResponse": {
                    "value": {
                      "message": "All notifications marked as read",
                      "data": {
                        "count": 10
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/{notificationId}": {
      "delete": {
        "summary": "Delete a notification",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "notificationId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Notification ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Notification deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "SuccessResponse": {
                    "value": {
                      "message": "Notification deleted",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Notification not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "NotFoundError": {
                    "$ref": "#/components/examples/ErrorExamples/NotFoundError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/preferences": {
      "get": {
        "summary": "Get user notification preferences",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Notification preferences",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "boolean"
                    },
                    "push": {
                      "type": "boolean"
                    },
                    "sms": {
                      "type": "boolean"
                    }
                  }
                },
                "examples": {
                  "SuccessResponse": {
                    "value": {
                      "email": true,
                      "push": false,
                      "sms": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update user notification preferences",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "boolean"
                  },
                  "push": {
                    "type": "boolean"
                  },
                  "sms": {
                    "type": "boolean"
                  }
                }
              },
              "examples": {
                "UpdatePreferences": {
                  "value": {
                    "email": true,
                    "push": true,
                    "sms": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification preferences updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "SuccessResponse": {
                    "value": {
                      "message": "Preferences updated",
                      "data": {
                        "email": true,
                        "push": true,
                        "sms": false
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "ValidationError": {
                    "$ref": "#/components/examples/ErrorExamples/ValidationError"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/notifications": {
      "get": {
        "summary": "Get user notification preferences",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Notification preferences",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "boolean"
                    },
                    "push": {
                      "type": "boolean"
                    },
                    "sms": {
                      "type": "boolean"
                    }
                  }
                },
                "examples": {
                  "SuccessResponse": {
                    "value": {
                      "email": true,
                      "push": false,
                      "sms": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update user notification preferences",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "boolean"
                  },
                  "push": {
                    "type": "boolean"
                  },
                  "sms": {
                    "type": "boolean"
                  }
                }
              },
              "examples": {
                "UpdatePreferences": {
                  "value": {
                    "email": true,
                    "push": true,
                    "sms": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification preferences updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "SuccessResponse": {
                    "value": {
                      "message": "Preferences updated",
                      "data": {
                        "email": true,
                        "push": true,
                        "sms": false
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "ValidationError": {
                    "$ref": "#/components/examples/ErrorExamples/ValidationError"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}